+++ src/dftbp/dftbplus/initprogram.F90
@@ -571,6 +571,9 @@ module dftbp_dftbplus_initprogram
     !> Dynamic polarisability at finite frequencies
     real(dp), allocatable :: dynKernelFreq(:)
 
+    !> Derivatives with respect to positions of external charges
+    logical :: isExtChargeDeriv
+
     !> Electric static polarisability
     real(dp), allocatable :: polarisability(:,:,:)
 
@@ -585,6 +588,20 @@ module dftbp_dftbplus_initprogram
 
     !> Derivatives of Mulliken charges with respect to perturbation
     real(dp), allocatable :: dqOut(:,:,:,:)
+
+    !> derivatives of atomic charges w.r.t. atom coordinates
+    real(dp), allocatable :: dQdX(:,:,:)
+
+    !> derivatives of atomic charges w.r.t. coordinates of MM atoms
+    real(dp), allocatable :: dQdXext(:,:,:)
+
+    !> number of MM atoms for whith the derivatives of atomic charges w.r.t. coordinates of those
+    !>   MM atoms shall be calculated
+    integer :: nExtChrgWRT
+
+    !> list of MM atoms for whith the derivatives of atomic charges w.r.t. coordinates of those MM
+    !>   atoms shall be calculated
+    integer, allocatable :: extChrgWRT(:)
 
     !> Use commands from socket communication to control the run
     logical :: tSocket
@@ -2523,6 +2540,8 @@ contains
         end if
       end if
 
+      this%isExtChargeDeriv = input%ctrl%perturbInp%isExtChargeDeriv
+
       this%isAtomCoordPerturb = input%ctrl%perturbInp%isAtomCoordPerturb
       if (this%isAtomCoordPerturb) then
         if (withMpi) then
@@ -2554,6 +2573,10 @@ contains
         if (this%tSpinOrbit) then
           call error("Spin-orbit coupling is not yet supported for coordinate derivative&
               & perturbations")
+        end if
+        if (this%boundaryCond%iBoundaryCondition /= boundaryCondsEnum%cluster) then
+          call error("Coordinate derivative perturbations not currently available for these&
+              & boundary condition")
         end if
       end if
       tRequireDerivator = tRequireDerivator .or. this%isAtomCoordPerturb